WorldObject ReOrg + local chat/emote broadcasting + ID redo
