+++ Source/ACE/ACE.csproj
@@ -104,16 +104,16 @@
     <Compile Include="Entity\Actions\LoopAction.cs" />
     <Compile Include="Entity\Actions\NestedActionQueue.cs" />
     <Compile Include="Entity\Adjacency.cs" />
-    <Compile Include="Entity\CollidableObject.cs" />
-    <Compile Include="Entity\Container.cs" />
-    <Compile Include="Entity\Creature.cs" />
-    <Compile Include="Entity\DebugObject.cs" />
-    <Compile Include="Entity\Door.cs" />
+    <Compile Include="Entity\WorldObjects\Collidable\CollidableObject.cs" />
+    <Compile Include="Entity\WorldObjects\Objects\Container.cs" />
+    <Compile Include="Entity\WorldObjects\Objects\Creature.cs" />
+    <Compile Include="Entity\WorldObjects\Collidable\Objects\DebugObject.cs" />
+    <Compile Include="Entity\WorldObjects\Usable\Objects\Door.cs" />
     <Compile Include="Entity\Events\DeathMessageArgs.cs" />
     <Compile Include="Entity\Events\ChatMessageArgs.cs" />
-    <Compile Include="Entity\Generator.cs" />
-    <Compile Include="Entity\Lifestone.cs" />
-    <Compile Include="Entity\Monster.cs" />
+    <Compile Include="Entity\WorldObjects\Objects\Generator.cs" />
+    <Compile Include="Entity\WorldObjects\Usable\Objects\Lifestone.cs" />
+    <Compile Include="Entity\WorldObjects\Objects\Monster.cs" />
     <Compile Include="Entity\MovementData.cs" />
     <Compile Include="Entity\Events\SpellCastArgs.cs" />
     <Compile Include="Entity\Landblock.cs" />
@@ -122,10 +122,10 @@
     <Compile Include="Entity\Model.cs" />
     <Compile Include="Entity\ModelPalette.cs" />
     <Compile Include="Entity\ModelTexture.cs" />
-    <Compile Include="Entity\Portal.cs" />
-    <Compile Include="Entity\SpellLikeEffect.cs" />
-    <Compile Include="Entity\SummonedPortal.cs" />
-    <Compile Include="Entity\UsableObject.cs" />
+    <Compile Include="Entity\WorldObjects\Collidable\Objects\Portal.cs" />
+    <Compile Include="Entity\WorldObjects\Objects\SpellLikeEffect.cs" />
+    <Compile Include="Entity\WorldObjects\Usable\Objects\SummonedPortal.cs" />
+    <Compile Include="Entity\WorldObjects\Usable\UsableObject.cs" />
     <Compile Include="Factories\CommonObjectFactory.cs" />
     <Compile Include="Factories\CorpseObjectFactory.cs" />
     <Compile Include="Factories\GeneratorFactory.cs" />
@@ -152,6 +152,7 @@
     <Compile Include="Network\GameAction\Actions\GameActionAddFriend.cs" />
     <Compile Include="Network\GameAction\Actions\GameActionAdvocateTeleport.cs" />
     <Compile Include="Network\GameAction\Actions\GameActionAllegianceUpdateRequest.cs" />
+    <Compile Include="Network\GameAction\Actions\GameActionSoulEmote.cs" />
     <Compile Include="Network\GameAction\Actions\GameActionStackableSplitToContainer.cs" />
     <Compile Include="Network\GameAction\Actions\GameActionChannelIndex.cs" />
     <Compile Include="Network\GameAction\Actions\GameActionChannelList.cs" />
@@ -188,8 +189,8 @@
     <Compile Include="Network\GameAction\Actions\GameActionTell.cs" />
     <Compile Include="Network\GameAction\GameActionAttribute.cs" />
     <Compile Include="Network\GameAction\GameActionType.cs" />
-    <Compile Include="Entity\Player.cs" />
-    <Compile Include="Entity\WorldObject.cs" />
+    <Compile Include="Entity\WorldObjects\Objects\Player.cs" />
+    <Compile Include="Entity\WorldObjects\WorldObject.cs" />
     <Compile Include="Managers\WorldManager.cs" />
     <Compile Include="Network\Enum\CharacterError.cs" />
     <Compile Include="Network\GameAction\Actions\GameActionMoveToState.cs" />
@@ -234,6 +235,7 @@
     <Compile Include="Network\GameMessages\Messages\GameMessagePutObjectIn3d.cs" />
     <Compile Include="Network\GameMessages\Messages\GameMessageSetStackSize.cs" />
     <Compile Include="Network\GameMessages\Messages\GameMessagePutObjectInContainer.cs" />
+    <Compile Include="Network\GameMessages\Messages\GameMessageSoulEmote.cs" />
     <Compile Include="Network\GameMessages\Messages\GameMessageUpdateInstanceId.cs" />
     <Compile Include="Network\Motion\MotionItem.cs" />
     <Compile Include="Network\GameMessages\Messages\GameMessageAutonomousPosition.cs" />
