+++ Source/ACE/Entity/WorldObjects/Collidable/CollidableObject.cs
