+++ Source/ACE/Entity/WorldObjects/Collidable/Objects/DebugObject.cs
