+++ Source/ACE/Entity/WorldObjects/Collidable/Objects/Portal.cs
