+++ Source/ACE/Entity/WorldObjects/Objects/Container.cs
