+++ Source/ACE/Entity/WorldObjects/Objects/Creature.cs
