+++ Source/ACE/Entity/WorldObjects/Objects/Generator.cs
