+++ Source/ACE/Entity/WorldObjects/Objects/SpellLikeEffect.cs
