+++ Source/ACE/Entity/WorldObjects/Usable/Objects/Door.cs
