+++ Source/ACE/Entity/WorldObjects/Usable/Objects/Lifestone.cs
