+++ Source/ACE/Entity/WorldObjects/Usable/Objects/SummonedPortal.cs
