+++ Source/ACE/Network/GameMessages/GameMessageOpcode.cs
@@ -5,7 +5,8 @@ namespace ACE.Network.GameMessages
         None                            = 0x0000,
         SetStackSize                    = 0x0197,
         PlayerKilled                    = 0x019E,
-        EmoteText                       = 0x01E2,
+        EmoteText                       = 0x01E0,
+        SoulEmote                       = 0x01E2,
         CreatureMessage                 = 0x02BB,
         PrivateUpdatePropertyInt        = 0x02CD,
         PublicUpdatePropertyInt         = 0x02CE,
