+++ Source/ACE/Network/GameMessages/Messages/GameMessageSoulEmote.cs
@@ -1,0 +1,13 @@
+﻿namespace ACE.Network.GameMessages.Messages
+{
+    public class GameMessageSoulEmote : GameMessage
+    {
+        public GameMessageSoulEmote(uint senderId, string senderName, string emoteText)
+            : base(GameMessageOpcode.SoulEmote, GameMessageGroup.Group09)
+        {
+            Writer.Write(senderId);
+            Writer.WriteString16L(senderName);
+            Writer.WriteString16L(emoteText);
+        }
+    }
+}
